requirement.txt
flask
passlib
pymongo
requests
pandas
xlrd
pyserial
pynmeagps
opencv-python
face_recognition
scikit-learn
pygrabber